perf(runtime): shrink DuxStr header to 16 bytes via int32_t len
Replacing int64_t len with int32_t len eliminates the 4-byte alignment
padding that followed refcount, shrinking sizeof(DuxStr) from 24 to 16:

  old: { i32 refcount, [4 pad], i64 len, ptr ext, char[] }  sizeof=24
  new: { i32 refcount,  i32 len,          ptr ext, char[] }  sizeof=16

Short strings (len <= DUXSTR_INLINE_MAX) now allocate malloc(16+len+1),
landing in the same glibc size-class as a pure-FAM struct (which also
has a 16-byte header).  Previously the 24-byte header pushed short strings
into a larger bin, explaining the speed gap.

Benchmark results vs two-alloc baseline (2 M ops, -O2, fair ->len reads):

  len=8:   hybrid (inline) ≈ fam-only   (+63–79%  vs two-alloc)
  len=40:  hybrid (inline)  ~ fam-only   (+74–77%  vs two-alloc)
  len=64:  hybrid (ext)      -3 to -7%   (two-alloc path, expected)
  len=200: hybrid (ext)      -3 to +2%   (beats fam-only by ~20%)

The remaining small gap at len=40 is the irreducible cost of the hybrid:
one extra store (ext=NULL) and one conditional branch in release.
Max string length is now 2 GiB (int32_t), which is sufficient for Dux.

Changes:
- duxrt.h: int32_t len field, updated sizeof comment
- str.c: cast to (int32_t) on store; explicit (int64_t) on public returns
- codegen.cpp: str_literal() emits { i32, i32, ptr, [N+1 x i8] } so the
               LLVM global layout stays aligned with the C struct (data
               at offset 16 instead of 24)
- benchmarks/str_alloc_bench.c: HybStr updated to int32_t len; all three
               variants now read ->len for a fair comparison

+++ src/runtime/duxrt.h
@@ -36,22 +36,27 @@ void  duxrt_print_double(double v);
  *   long strings  (len >  DUXSTR_INLINE_MAX):
  *       malloc(sizeof(DuxStr)) for the header    — ext points to heap buffer
  *
- * sizeof(DuxStr) = 24 bytes on LP64:
+ * sizeof(DuxStr) = 16 bytes on LP64:
  *   offset  0: int32_t refcount
- *   offset  4: [4 bytes padding]
- *   offset  8: int64_t len
- *   offset 16: char*   ext     (NULL = inline)
- *   offset 24: char    data[]  (FAM — inline data for short strings)
+ *   offset  4: int32_t len      (max 2 GiB; int32_t avoids padding after refcount)
+ *   offset  8: char*   ext      (NULL = inline)
+ *   offset 16: char    data[]   (FAM — inline data for short strings)
  *
- * LLVM literal globals use { i32, i64, ptr, [N+1 x i8] } which produces the
+ * Using int32_t for len instead of int64_t eliminates 4 bytes of alignment
+ * padding, shrinking the header from 24 to 16 bytes.  Short strings then
+ * allocate malloc(16+len+1), landing in the same allocator size-class as a
+ * pure-FAM struct — eliminating the bin-boundary penalty seen with 24-byte
+ * headers.
+ *
+ * LLVM literal globals use { i32, i32, ptr, [N+1 x i8] } which produces the
  * same offsets, so immortal literals always have ext=null and the string bytes
- * at offset 24 — duxrt_str_cstr() returns s->data for them correctly.
+ * at offset 16 — duxrt_str_cstr() returns s->data for them correctly.
  */
 #define DUXSTR_INLINE_MAX  63   /* strings <= 63 bytes are stored inline (FAM) */
 
 typedef struct DuxStr {
     int32_t  refcount;
-    int64_t  len;
+    int32_t  len;    /* int32_t: no padding after refcount → header = 16 bytes */
     char*    ext;    /* NULL  → data inline in FAM below (short strings)
                         non-NULL → separate heap buffer    (long strings)  */
     char     data[]; /* inline storage — only valid when ext == NULL */
